requirements.txt
asgiref==3.11.1
crispy-bootstrap5==2026.3
Django==5.2.12
django-crispy-forms==2.6
pillow==12.1.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
